--HG--
branch : release


Former-commit-id: 95135c89f5d8f6e76869531cb7c8c30bf3d4cced
Former-commit-id: d56ee6f2fd71e70fcaedf649ddb959843c442334
Former-commit-id: 7a2217f37ab2af976ae03129dcba84fc59625668

+++ src/main/java/teammates/ui/controller/InstructorFeedbackEditSaveAction.java
@@ -54,7 +54,7 @@ public class InstructorFeedbackEditSaveAction extends Action {
 				getRequestParamValue(Const.ParamsNames.FEEDBACK_SESSION_ENDTIME));		
 		String paramTimeZone = getRequestParamValue(Const.ParamsNames.FEEDBACK_SESSION_TIMEZONE);
 		if (paramTimeZone != null) {
-			newSession.timeZone = Integer.parseInt(paramTimeZone);
+			newSession.timeZone = Double.parseDouble(paramTimeZone);
 		}
 		String paramGracePeriod = getRequestParamValue(Const.ParamsNames.FEEDBACK_SESSION_GRACEPERIOD);
 		if (paramGracePeriod != null) {
